backend/requirements.txt
django>=5.1,<6.0
django-ninja>=1.3,<2.0
psycopg2-binary>=2.9,<3.0
django-cors-headers>=4.3,<5.0
python-decouple>=3.8,<4.0
gunicorn>=22.0,<23.0
dj-database-url>=2.1,<3.0
whitenoise>=6.6,<7.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
pytest-cov>=5.0,<6.0
factory-boy>=3.3,<4.0
